+++ Game/Source/LightTimer.h
@@ -12,7 +12,7 @@ public:
 
 	void Start();
 	uint32 Read() const;
-	float ReadSec() const;
+	double ReadSec() const;
 
 private:
 	uint32 startTime;
