+++ Game/Source/PerfTimer.cpp
@@ -32,7 +32,7 @@ uint64 PerfTimer::ReadTicks() const
 	return SDL_GetPerformanceCounter() - startTime;
 }
 
-float PerfTimer::ReadSec() const
+double PerfTimer::ReadSec() const
 {
-	return float(SDL_GetTicks() - started_at) / 1000.0f;
+	return (SDL_GetTicks() - started_at) / 1000.0;
 }
